usb: pd_engine: Fix memory leaks in probe and destroy
dev_set_name() allocates memory to the provided device instance's kobject
in order to store a name. usbpd_create() returns -EPROBE_DEFER the first
time it's called, which causes it to go down the error path releasing
resources, but because device_del() never calls kobject_put() on the
kobject (it deletes it with kobject_del()), the memory allocated to the
name is never freed. Only kobject_put() frees memory allocated to the
kobject's name.

Additionally, memory allocated for the private data pointer in the
device struct (pd->dev.p) is leaked upon error as well.

Now, replacing device_del() with device_unregister() fixes both memory
leaks (because device_unregister() calls kobject_put(), which frees the
name, and the device struct is totally freed), but because this driver's
device doesn't have a release method, device_unregister() leaves us with a
nasty warning:
"Device 'usbpd0' does not have a release() function, it is broken and must be fixed."

Therefore, in order to fix the memory leaks without angering anything,
add an empty release function to the device instance so that
device_unregister() can be used, and explicitly free the memory allocated
for the name with kfree_const() when init fails before device_unregister()
can be used. kfree_const is used because it's what kobject_put() uses to
free the names it allocates. Note that the kfree_const() cannot live inside
the error labels to avoid a double free with device_unregister().

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: SantiagoPintos <santiagoopintos@hotmail.com>